+++ applications/platform/designer/src/ovdAssert.h
@@ -62,7 +62,7 @@ public:
 #define OV_EXCEPTION_D(description, type) \
 do { \
 	m_kernelCtx.getErrorManager().pushErrorAtLocation(type, static_cast<const OpenViBE::ErrorStream&>(OpenViBE::ErrorStream() << description).str().c_str(), __FILE__, __LINE__ ); \
-	m_kernelCtx.getLogManager() << OpenViBE::Kernel::LogLevel_Fatal << "[Error description] = " << description << "; [Error code] = " << uint32_t((type)) << "\n"; \
+	m_kernelCtx.getLogManager() << OpenViBE::Kernel::LogLevel_Fatal << "[Error description] = " << description << "; [Error code] = " << size_t((type)) << "\n"; \
 	throw DesignerException(m_kernelCtx.getErrorManager()); \
 } while(0)
